refactor: move minikube cluster nginx code to a dedicated file
For better modularity and readability.
