+++ core/src/plugins/kubernetes/local/minikube.ts
@@ -7,17 +7,7 @@
  */
 
 import { ChildProcessError } from "../../../exceptions.js"
-import { Log } from "../../../logger/log-entry.js"
-import type { DeployState } from "../../../types/service.js"
-import { exec } from "../../../util/util.js"
-import chalk from "chalk"
-
-interface minikubeAddons {
-  [key: string]: {
-    Profile: string
-    Status: string
-  }
-}
+import type { exec } from "../../../util/util.js"
 
 /**
  * Automatically set docker environment variables for minikube
@@ -44,30 +34,4 @@ export async function setMinikubeDockerEnv() {
       process.env[matched[1]] = matched[2]
     }
   }
-}
-
-export async function minikubeNginxStatus(log: Log): Promise<DeployState> {
-  const result = await exec("minikube", ["addons", "list", "-o=json"])
-  const minikubeAddons = JSON.parse(result.stdout) as minikubeAddons
-  const addonEnabled = minikubeAddons.ingress.Status === "enabled"
-  log.debug(chalk.yellow(`Status of minikube ingress controller addon: ${addonEnabled ? "enabled" : "disabled"}`))
-  return addonEnabled ? "ready" : "missing"
-}
-
-export async function minikubeNginxInstall(log: Log) {
-  const status = await minikubeNginxStatus(log)
-  if (status === "ready") {
-    return
-  }
-  log.info("Enabling minikube ingress controller addon")
-  await exec("minikube", ["addons", "enable", "ingress"])
-}
-
-export async function minikubeNginxUninstall(log: Log) {
-  const status = await minikubeNginxStatus(log)
-  if (status === "missing") {
-    return
-  }
-  log.info("Disabling minikube ingress controller addon")
-  await exec("minikube", ["addons", "disable", "ingress"])
 }
